+++ StreamChatSwiftUITests/Tests/Utils/ViewFactory_Tests.swift
@@ -912,6 +912,28 @@ class ViewFactory_Tests: StreamChatTestCase {
         // Then
         XCTAssert(view is LoadingView)
     }
+    
+    func test_viewFactory_makeComposerPollView() {
+        // Given
+        let viewFactory = DefaultViewFactory.shared
+        
+        // When
+        let view = viewFactory.makeComposerPollView()
+        
+        // Then
+        XCTAssert(view is ComposerPollView)
+    }
+    
+    func test_viewFactory_makePollView() {
+        // Given
+        let viewFactory = DefaultViewFactory.shared
+        
+        // When
+        let view = viewFactory.makePollView(message: .mock(), poll: Poll.mock(), isFirst: true)
+        
+        // Then
+        XCTAssert(view is PollAttachmentView<DefaultViewFactory>)
+    }
 }
 
 extension ChannelAction: Equatable {
